Template is now rendered with Jinja (by Flask) and served by Flask.
IP/PORT need now to be specied in server.py and should correspond
to the IP of the websocket server and the PORT specified in
websocket_server.py.
Moved stuff around to get the Jinja templating right.
Changed filenames for clearer overview.
Added README.
