+++ server.py
@@ -1,66 +1,35 @@
-import asyncio
-import websockets
-import json
-import time
-import hashlib
-from watchdog.observers import Observer
-from watchdog.events import FileSystemEventHandler
-from collections import defaultdict
+from flask import Flask, request, url_for, render_template
+from passlib.hash import pbkdf2_sha256
+import utils
 
-FILE = "status_kfet.json"
+app = Flask(__name__)
 
-def json_error():
-    print("JSON error : attribute not found")
-    return
+IP="127.0.0.1"
+# TODO set the PORT specified in websocket_server.py
+PORT="8080"
+# TODO update the password
+secret_hash = pbkdf2_sha256.hash("somepassword")
 
-# load kfet status from json file f
-def get_status(f):
-    with open(f) as fd:
-        kfet = defaultdict(json_error, json.load(fd))
-    return kfet["status"]
+@app.route('/')
+def root():
+	return render_template("kfet.html", IP=IP, PORT=PORT)
 
-class UpdateHandler(FileSystemEventHandler):
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+	if request.method == 'POST':
+		if check_auth(request.form['password']):
+			utils.toggle_state()
+			return "Toggled state"
+		return "Invalid password"
 
-    def __init__(self, f):
-        super().__init__()
-        self.connected = []
+	else:
+		return '''
+			<form method="post">
+            <p><input type=text name=password>
+			<p><input type=submit value=Log>
+        </form>
+    '''
 
-    async def send(self, msg):
-        for i, ws in enumerate(self.connected):
-            try : 
-                await ws.send(msg)
-            except websockets.ConnectionClosed:
-                self.connected[i] = False
-        
-        # remove all disconnected webscockets 
-        self.connected = list(filter((lambda connected: connected), self.connected))
+def	check_auth(password):
+	return pbkdf2_sha256.verify(password, secret_hash)
 
-    def on_modified(self, event):
-        status = get_status(FILE)
- 
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        loop.run_until_complete(self.send(status))
-
-
-event_handler = UpdateHandler(FILE)
-
-async def update(websocket, path):
-    event_handler.connected.append(websocket)
-
-    # init connection by sending the current
-    # status
-    await websocket.send(get_status(FILE))
-
-    while True:
-        await asyncio.sleep(1) 
-
-websocket = websockets.serve(update, 'localhost', 7800)
-loop = asyncio.get_event_loop()
-observer = Observer()
-observer.schedule(event_handler, "./")
-observer.start()
-
-asyncio.set_event_loop(loop)
-loop.run_until_complete(websocket)
-loop.run_forever()
